examples/droid/webui/requirements.txt
# WebUI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.0.0
pyyaml>=6.0
python-multipart>=0.0.6
jinja2>=3.1.0
websockets>=12.0


